requirements.txt
decouple
bandit
black
coverage
django-naomi
isort
model-mommy
pre-commit
prospector[with-vulture]
safety
django
celery[redis]
django-model-utils
django-webpack-loader
django-js-reverse
django-import-export
django-ses
python-decouple
psycopg2
brotlipy
django-log-request-id
dj-database-url
gunicorn
whitenoise
psutil
ipython
sentry-sdk
djangorestframework
django-rest-auth
requests
social-auth-core
django-background-tasks==1.2.5
django-compat==1.0.15
pytz==2020.1
six==1.15.0
